Move map layers state to recoil
